+++ src/Trash/Radarr/CustomFormat/Guide/CustomFormatData.cs
@@ -1,8 +1,0 @@
-﻿namespace Trash.Radarr.CustomFormat.Guide
-{
-    public class CustomFormatData
-    {
-        public int? Score { get; set; } = null;
-        public string Json { get; set; } = "";
-    }
-}