+++ src/Trash/Radarr/CustomFormat/Guide/IRadarrGuideService.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace Trash.Radarr.CustomFormat.Guide
+{
+    public interface IRadarrGuideService
+    {
+        Task<IEnumerable<string>> GetCustomFormatJson();
+    }
+}